fix: AnswerItem type 위치 변경
