+++ src/hooks/api/apply/useSingleAnswer.ts
@@ -1,16 +1,13 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError, type AxiosResponse } from 'axios';
 import { getSingleAnswer } from '@/apis';
-export type AnswerItem = {
-  applicationId: number;
-  name: string;
-  answer: string;
-  fileName?: string;
-};
+import { AnswerItem } from '@/types/apply';
+
 type Props = {
   type: 'TEXT' | 'LONG_TEXT' | 'FILE';
   answers: AnswerItem[];
 };
+
 export function useSingleAnswer(questionId: number, token: string) {
   return useQuery<
     unknown,
